abego-commons-base/src/main/java/org/abego/commons/pair/IntPairDefault.java
/*
 * MIT License
 *
 * Copyright (c) 2020 Udo Borkowski, (ub@abego.org)
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */
package org.abego.commons.pair;

import org.eclipse.jdt.annotation.Nullable;

import java.util.Objects;

public final class IntPairDefault implements IntPair {

    private final int first;
    private final int second;

    IntPairDefault(int first, int second) {
        this.first = first;
        this.second = second;
    }

    public static IntPair newIntPair(int first, int second) {
        return new IntPairDefault(first, second);
    }

    public static IntPairDefault of(int first, int second) {
        return new IntPairDefault(first, second);
    }

    @Override
    public int first() {
        return first;
    }

    @Override
    public int second() {
        return second;
    }

    @Override
    public boolean equals(@Nullable Object o) {
        if (this == o) return true;
        if (!(o instanceof IntPair)) return false;
        IntPair intPair = (IntPair) o;
        return first == intPair.first() && second == intPair.second();
    }

    @Override
    public int hashCode() {
        return Objects.hash(first, second);
    }

    @Override
    public String toString() {
        return String.format("(%d,%d)", first, second); //NON-NLS
    }
}